fix new device context

+++ device_context.go
@@ -6,7 +6,7 @@ import "C"
 
 func NewDeviceContext(deviceType DeviceTypeFlag) *BufferRef {
 	br := newBufferRef()
-	ret := int(C.av_hwdevice_ctx_create(&br.c, deviceType, nil, nil, 0))
+	ret := int(C.av_hwdevice_ctx_create(&br.c, uint32(deviceType), nil, nil, 0))
 	if ret < 0 {
 		return nil
 	} else {
